requirements.txt
httpx
python-dotenv
mcp-server
pydantic 